applications: nrf5340_audio: Implement scanner for conn
Part of the architecture rewrite.

OCT-2576
