+++ applications/nrf5340_audio/src/bluetooth/le_audio.h
@@ -11,20 +11,7 @@
 #include <zephyr/bluetooth/audio/audio.h>
 #include <audio_defines.h>
 
-#define DEVICE_NAME_PEER CONFIG_BT_DEVICE_NAME
-#define DEVICE_NAME_PEER_LEN (sizeof(DEVICE_NAME_PEER) - 1)
-
 #define LE_AUDIO_SDU_SIZE_OCTETS(bitrate) (bitrate / (1000000 / CONFIG_AUDIO_FRAME_DURATION_US) / 8)
-
-#if (CONFIG_SCAN_MODE_ACTIVE)
-#define NRF5340_AUDIO_GATEWAY_SCAN_TYPE BT_LE_SCAN_TYPE_ACTIVE
-#define NRF5340_AUDIO_GATEWAY_SCAN_PARAMS BT_LE_SCAN_ACTIVE
-#elif (CONFIG_SCAN_MODE_PASSIVE)
-#define NRF5340_AUDIO_GATEWAY_SCAN_TYPE BT_LE_SCAN_TYPE_PASSIVE
-#define NRF5340_AUDIO_GATEWAY_SCAN_PARAMS BT_LE_SCAN_PASSIVE
-#else
-#error "Please select either CONFIG_SCAN_MODE_ACTIVE or CONFIG_SCAN_MODE_PASSIVE"
-#endif
 
 #if (CONFIG_AUDIO_SAMPLE_RATE_48000_HZ)
 #define BT_AUDIO_CODEC_CONFIG_FREQ BT_CODEC_CONFIG_LC3_FREQ_48KHZ
@@ -129,6 +116,21 @@ typedef void (*le_audio_receive_cb)(const uint8_t *const data, size_t size, bool
  */
 typedef void (*le_audio_timestamp_cb)(uint32_t timestamp, bool adjust);
 
+/**
+ * @brief Callback for starting to scan for Bluetooth LE advertisements
+ */
+typedef void (*le_audio_scan_start_cb)(void);
+
+/**
+ * @brief Callback for notifying about nonvalid ISO configs for a connection
+ *
+ * @note  Nonvalid configs means that the local device does not support the
+ *        configurations that the remote device supports
+ *
+ * @param conn          The remote device connection
+ */
+typedef void (*le_audio_nonvalid_iso_cfgs_cb)(struct bt_conn *conn);
+
 enum le_audio_user_defined_action {
 	LE_AUDIO_USER_DEFINED_ACTION_1,
 	LE_AUDIO_USER_DEFINED_ACTION_2,
@@ -177,8 +179,17 @@ int le_audio_config_get(uint32_t *bitrate, uint32_t *sampling_rate, uint32_t *pr
 void le_audio_conn_set(struct bt_conn *conn);
 
 /**
+ * @brief	Notify about disconnected connection, used by CIS
+ *
+ * @param[in]	conn	The connection pointer
+ */
+void le_audio_conn_disconnected(struct bt_conn *conn);
+
+/**
  * @brief	Set pointer to extended advertisement, used by BIS
+ *
  * @note	Will also start the broadcast_source when used with BIS
+ *
  * @param[in]	conn	The connection pointer
  *
  * @return      0 for success,
@@ -244,12 +255,14 @@ int le_audio_send(struct encoded_audio enc_audio);
 /**
  * @brief Enable Bluetooth LE Audio
  *
- * @param recv_cb	Callback for receiving Bluetooth LE Audio data
- * @param timestamp_cb	Callback for using timestamp
+ * @param recv_cb		Callback for receiving Bluetooth LE Audio data
+ * @param timestmp_cb		Callback for using timestamp
+ * @param nonvalid_cfgs_cb	Callback for notifying about nonvalid cfgs for a conn
  *
  * @return		0 for success, error otherwise
  */
-int le_audio_enable(le_audio_receive_cb recv_cb, le_audio_timestamp_cb timestamp_cb);
+int le_audio_enable(le_audio_receive_cb recv_cb, le_audio_timestamp_cb timestmp_cb,
+		    le_audio_nonvalid_iso_cfgs_cb nonvalid_cfgs_cb);
 
 /**
  * @brief Disable Bluetooth LE Audio
