+++ applications/nrf5340_audio/src/bluetooth/le_audio_bis_headset.c
@@ -626,7 +626,8 @@ int le_audio_send(struct encoded_audio enc_audio)
 	return -ENXIO;
 }
 
-int le_audio_enable(le_audio_receive_cb recv_cb, le_audio_timestamp_cb timestmp_cb)
+int le_audio_enable(le_audio_receive_cb recv_cb, le_audio_timestamp_cb timestmp_cb,
+		    le_audio_nonvalid_iso_cfgs_cb nonvalid_cfgs_cb)
 {
 	int ret;
